Add support for external domain to certificate
Config option DOMAIN will allow adding one subjectAltName to the
self-signed certificate.

This will allow (external) users of the certificate, e.g. Wallets, to
validate it according to the supplied domain.

The option will default to "localhost" in the example config so current
behaviour remains unchanged.

It also changes the commonName (CN) of the certificate to
"c-lightning-rest" for easier identification amongst certificates.